+++ src/eth/executor/mod.rs
@@ -24,13 +24,13 @@ use tracing::debug_span;
 use tracing::info_span;
 pub use types::AccountChanges;
 pub use types::AccountOriginalsReader;
-pub use types::ChangedOrOriginalValue;
-pub use types::ChangedValue;
+pub use types::Changes;
 pub use types::Complete;
-pub use types::ExecutionChanges;
+pub use types::CompleteValue;
 pub use types::ExecutionResult;
 pub use types::ExecutorError;
 pub use types::Incomplete;
+pub use types::IncompleteValue;
 pub use types::RevertReason;
 pub use types::Stage;
 pub use types::TransactionExecution;
