+++ src/eth/executor/types/mod.rs
@@ -1,19 +1,18 @@
-mod account_changes;
+mod changes;
 mod error;
-mod execution_changes;
 mod execution_result;
 mod task;
 mod transaction_execution;
 
-pub use account_changes::AccountChanges;
-pub use account_changes::ChangedOrOriginalValue;
-pub use account_changes::ChangedValue;
+pub use changes::AccountChanges;
+pub use changes::AccountOriginalsReader;
+pub use changes::Changes;
+pub use changes::Complete;
+pub use changes::CompleteValue;
+pub use changes::Incomplete;
+pub use changes::IncompleteValue;
+pub use changes::Stage;
 pub use error::ExecutorError;
-pub use execution_changes::AccountOriginalsReader;
-pub use execution_changes::Complete;
-pub use execution_changes::ExecutionChanges;
-pub use execution_changes::Incomplete;
-pub use execution_changes::Stage;
 pub use execution_result::ExecutionResult;
 pub use execution_result::RevertReason;
 pub use task::EvmRoute;
