+++ src/eth/follower/importer/importers/replication.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use async_trait::async_trait;
 
-use crate::eth::executor::ExecutionChanges;
+use crate::eth::executor::Changes;
 use crate::eth::executor::Incomplete;
 use crate::eth::follower::importer::importers::ImportData;
 use crate::eth::follower::importer::importers::ImporterWorker;
@@ -27,7 +27,7 @@ impl ImportData for <ReplicationWorker as ImporterWorker>::DataType {
 
 #[async_trait]
 impl ImporterWorker for ReplicationWorker {
-    type DataType = (Block, ExecutionChanges<Incomplete>);
+    type DataType = (Block, Changes<Incomplete>);
 
     async fn import(&self, (block, changes): Self::DataType) -> anyhow::Result<usize> {
         tracing::info!(block_number = %block.number(), "received block with changes");
